(be): Clean up query strings in fe post routes
Notes:
- Use sequelize replacements to avoid sql injection when using raw sql
- Raw sql necessary to make subqueries
- Subqueries necessary to filter posts where category matches, but return
  all categories of those posts. With simple WHERE, categories would be
  filtered as well after the join
- Raw query unneccessary on user query as only one user on post
